+++ src/main/java/ru/akvine/istochnik/providers/converters/IntegerConvertersProvider.java
@@ -5,8 +5,8 @@ import ru.akvine.istochnik.services.converters.integer.IntegerConverter;
 
 import java.util.Map;
 
-public record IntegerConvertersProvider(Map<ConverterType, IntegerConverter<Integer, Double>> converters) {
-    public IntegerConverter<Integer, Double> getConverter(ConverterType type) {
+public record IntegerConvertersProvider(Map<ConverterType, IntegerConverter<Long, Double>> converters) {
+    public IntegerConverter<Long, Double> getConverter(ConverterType type) {
         return converters.get(type);
     }
 }
