+++ src/main/java/ru/akvine/istochnik/services/converters/integer/PowIntegerConverter.java
@@ -6,10 +6,10 @@ import ru.akvine.istochnik.enums.ConverterType;
 import java.util.List;
 
 @Service
-public class PowIntegerConverter extends IntegerConverter<Integer, Double> {
+public class PowIntegerConverter extends IntegerConverter<Long, Double> {
     @Override
-    public List<Integer> convert(List<Integer> input, Double[] argument) {
-        return input.stream().map(value -> (int) Math.pow(value, argument[0])).toList();
+    public List<Long> convert(List<Long> input, Double[] argument) {
+        return input.stream().map(value -> (long) Math.pow(value, argument[0])).toList();
     }
 
     @Override
